requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
packaging==25.0
python-docx==1.1.2
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
